# DriveSistem/admin_sistem/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Token Authentication",
    "description": "API для администрирования системы",
    "version": "1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Healthcheck",
        "operationId": "healthcheck_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    },
    "/token": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Получение токена",
        "operationId": "token_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_token_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Вход в аккаунт(Админестратор)",
        "operationId": "login_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/drivers/": {
      "get": {
        "tags": [
          "Водители"
        ],
        "summary": "Получение всех водителей",
        "operationId": "get_drivers_admin_drivers__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {

                  },
                  "type": "array",
                  "title": "Response Get Drivers Admin Drivers  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/admin/driver/{id}/": {
      "delete": {
        "tags": [
          "Водители"
        ],
        "summary": "Удаление водителя по id",
        "operationId": "delete_driver_admin_driver__id___delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Водители"
        ],
        "summary": "Получение водителя по id",
        "operationId": "get_driver_admin_driver__id___get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Водители"
        ],
        "summary": "Изменение водителя по id",
        "operationId": "change_driver_admin_driver__id___post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Driver"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/passengers/": {
      "get": {
        "tags": [
          "Пассажиры"
        ],
        "summary": "Получение всех пассажиров",
        "operationId": "get_passengers_admin_passengers__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {

                  },
                  "type": "array",
                  "title": "Response Get Passengers Admin Passengers  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/admin/passenger/{id}/": {
      "delete": {
        "tags": [
          "Пассажиры"
        ],
        "summary": "Удаление пассажира по id",
        "operationId": "delete_passenger_admin_passenger__id___delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Пассажиры"
        ],
        "summary": "Получение пассажира по id",
        "operationId": "get_passenger_admin_passenger__id___get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Пассажиры"
        ],
        "summary": "Изменение пассажира по id",
        "operationId": "change_passenger_admin_passenger__id___post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Passenger"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/trips/": {
      "get": {
        "tags": [
          "Поездки"
        ],
        "summary": "Получение всех поездок",
        "operationId": "get_trips_admin_trips__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {

                  },
                  "type": "array",
                  "title": "Response Get Trips Admin Trips  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/admin/trip/{id}/": {
      "delete": {
        "tags": [
          "Поездки"
        ],
        "summary": "Удаление поездки по id",
        "operationId": "delete_trip_admin_trip__id___delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Поездки"
        ],
        "summary": "Получение поездки по id",
        "operationId": "get_trip_admin_trip__id___get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Поездки"
        ],
        "summary": "Изменение поездки по id",
        "operationId": "change_trip_admin_trip__id___post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trip"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Account": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Account"
      },
      "Body_token_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_token_token_post"
      },
      "Driver": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "first_name": {
            "type": "string",
            "pattern": "^[A-Za-zА-Яа-яЁё]{2,50}$",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "pattern": "^[A-Za-zА-Яа-яЁё]{2,50}$",
            "title": "Last Name"
          },
          "patronymic": {
            "type": "string",
            "pattern": "^[A-Za-zА-Яа-яЁё]{2,50}$",
            "title": "Patronymic"
          },
          "phone_number": {
            "type": "string",
            "pattern": "^\\+?\\d{10,15}$",
            "title": "Phone Number"
          },
          "requisites": {
            "type": "string",
            "title": "Requisites"
          },
          "car_number": {
            "type": "string",
            "title": "Car Number"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "first_name",
          "last_name",
          "patronymic",
          "phone_number",
          "requisites",
          "car_number"
        ],
        "title": "Driver"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Passenger": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "first_name": {
            "type": "string",
            "pattern": "^[A-Za-zА-Яа-яЁё]{2,50}$",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "pattern": "^[A-Za-zА-Яа-яЁё]{2,50}$",
            "title": "Last Name"
          },
          "patronymic": {
            "type": "string",
            "pattern": "^[A-Za-zА-Яа-яЁё]{2,50}$",
            "title": "Patronymic"
          },
          "phone_number": {
            "type": "string",
            "pattern": "^\\+?\\d{10,15}$",
            "title": "Phone Number"
          },
          "home_address": {
            "type": "string",
            "title": "Home Address"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "first_name",
          "last_name",
          "patronymic",
          "phone_number",
          "home_address"
        ],
        "title": "Passenger"
      },
      "Trip": {
        "properties": {
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "trip_info": {
            "$ref": "#/components/schemas/TripDescription"
          },
          "passengers": {
            "items": {
              "$ref": "#/components/schemas/Passenger"
            },
            "type": "array",
            "title": "Passengers"
          },
          "driver": {
            "$ref": "#/components/schemas/Driver"
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "trip_info",
          "passengers",
          "driver"
        ],
        "title": "Trip"
      },
      "TripDescription": {
        "properties": {
          "info": {
            "type": "string",
            "title": "Info"
          }
        },
        "type": "object",
        "required": [
          "info"
        ],
        "title": "TripDescription"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}